more checkers, more changes

+++ src/gb-lib/location/locationimpl.cpp
@@ -8,8 +8,9 @@ LocationImpl::LocationImpl(LocationByteUP lower, LocationByteUP upper)
 uint8_t LocationImpl::getByte() const { return **lower_; }
 
 uint16_t LocationImpl::getWord() const {
-  uint16_t result = (static_cast<uint16_t>(**upper_) << BYTE_SIZE) |
-                    static_cast<uint16_t>(**lower_);
+  uint16_t result =
+      static_cast<uint16_t>(static_cast<uint16_t>(**upper_) << BYTE_SIZE) |
+      static_cast<uint16_t>(**lower_);
   return result;
 }
 
@@ -17,7 +18,7 @@ void LocationImpl::setByte(uint8_t value) { *lower_ << value; }
 
 void LocationImpl::setWord(uint16_t value) {
   uint8_t lowerValue = value & LOWER_MASK;
-  uint8_t upperValue = (value & UPPER_MASK) >> BYTE_SIZE;
+  uint8_t upperValue = static_cast<uint16_t>(value & UPPER_MASK) >> BYTE_SIZE;
   *lower_ << lowerValue;
   *upper_ << upperValue;
 }
